refactor: updated files in pages/api/webhooks folder to Typescript

+++ pages/api/webhooks/webhookDelOrg.ts
@@ -1,7 +1,7 @@
 import { Webhook } from "svix";
-import { buffer } from "@/lib/buffer";
 import { getSupabaseClient } from "@/lib/supabase";
 import { propelauth } from "@/lib/propelauth";
+import type { NextApiRequest, NextApiResponse } from 'next'
 
 export const config = {
     api: {
@@ -11,25 +11,40 @@ export const config = {
 
 const secret = process.env.SVIX_WEBHOOK_DEL_ORG;
 
-export default async function POST(req, res) {
+type ResponseData = {
+    message?: string;
+    error?: string;
+};
+
+export default async function POST(request: NextApiRequest, response: NextApiResponse<ResponseData>) {
 
     console.log("Webhook received! Verifying...");
 
-    if(req.method !== "POST"){
-        res.status(405).json({
+    //check if the method is POST
+    if(request.method !== "POST"){
+        response.status(405).json({
             error: "Method not allowed"
         });
     }
 
-    const payload = (await buffer(req)).toString();
-    const headers = req.headers;
+    //check if the secret is defined
+    if(typeof secret === "undefined"){
+        return response.status(500).json({
+            error: "Internal server error"
+        });
+    }
 
+    //read the request
+    const payload = request.body;
+    const headers: any = request.headers;
+
+    //verify the webhook
     const wh = new Webhook(secret);
-    let msg;
+    let msg: any;
     try {
         msg = wh.verify(payload, headers);
     } catch (err) {
-        res.status(400).json({});
+        return response.status(400).json({});
     }
 
     console.log("Webhook verified! Starting to process...");
@@ -43,6 +58,13 @@ export default async function POST(req, res) {
     //fetch the users from the database
     const { data: users, error: error1 } = await supabase.from("user_table").select("user_id").eq("org_id", org_id);
 
+    //check for errors
+    if(!users){
+        return response.status(500).json({
+            error: "Internal server error"
+        });
+    }
+
     //delete from the database
     const { error: error2 } = await supabase.from("org_table").delete().eq("org_id", org_id);
 
@@ -54,12 +76,12 @@ export default async function POST(req, res) {
     };
 
     //check for errors
-    if(error){
-        console.log("Error inserting data into the database: ", error);
-        res.status(500).json({
-            error: "Error inserting data into the database"
+    if(error1 || error2){
+        console.log("Error in the delete org webhook:", error1, error2);
+        return response.status(500).json({
+            error: "Internal server error"
         });
     };
 
-    res.json({message: "Webhook processed successfully!"});
+    return response.json({message: "Webhook processed successfully!"});
 };
